requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-json-logger==2.0.7

# Oracle driver + синхронный SQLAlchemy
oracledb==2.1.2
sqlalchemy==2.0.25          # ← без [asyncio]
alembic==1.13.1
python-dotenv==1.0.0
defusedxml>=0.7.1